+++ react-library/src/components/stencil-generated/index.ts
@@ -1,0 +1,62 @@
+/* eslint-disable */
+/* tslint:disable */
+/* auto-generated react proxies */
+import { createReactComponent } from './react-component-lib';
+
+import type { JSX } from '@sbb-esta/lyne-components';
+
+import { defineCustomElements } from '@sbb-esta/lyne-components/loader';
+
+defineCustomElements();
+export const SbbAccordion = /*@__PURE__*/createReactComponent<JSX.SbbAccordion, HTMLSbbAccordionElement>('sbb-accordion');
+export const SbbAccordionItem = /*@__PURE__*/createReactComponent<JSX.SbbAccordionItem, HTMLSbbAccordionItemElement>('sbb-accordion-item');
+export const SbbAutocomplete = /*@__PURE__*/createReactComponent<JSX.SbbAutocomplete, HTMLSbbAutocompleteElement>('sbb-autocomplete');
+export const SbbAutocompleteItem = /*@__PURE__*/createReactComponent<JSX.SbbAutocompleteItem, HTMLSbbAutocompleteItemElement>('sbb-autocomplete-item');
+export const SbbButton = /*@__PURE__*/createReactComponent<JSX.SbbButton, HTMLSbbButtonElement>('sbb-button');
+export const SbbCardBadge = /*@__PURE__*/createReactComponent<JSX.SbbCardBadge, HTMLSbbCardBadgeElement>('sbb-card-badge');
+export const SbbCardProduct = /*@__PURE__*/createReactComponent<JSX.SbbCardProduct, HTMLSbbCardProductElement>('sbb-card-product');
+export const SbbClock = /*@__PURE__*/createReactComponent<JSX.SbbClock, HTMLSbbClockElement>('sbb-clock');
+export const SbbDivider = /*@__PURE__*/createReactComponent<JSX.SbbDivider, HTMLSbbDividerElement>('sbb-divider');
+export const SbbFooter = /*@__PURE__*/createReactComponent<JSX.SbbFooter, HTMLSbbFooterElement>('sbb-footer');
+export const SbbFormError = /*@__PURE__*/createReactComponent<JSX.SbbFormError, HTMLSbbFormErrorElement>('sbb-form-error');
+export const SbbFormField = /*@__PURE__*/createReactComponent<JSX.SbbFormField, HTMLSbbFormFieldElement>('sbb-form-field');
+export const SbbGrid = /*@__PURE__*/createReactComponent<JSX.SbbGrid, HTMLSbbGridElement>('sbb-grid');
+export const SbbIcon = /*@__PURE__*/createReactComponent<JSX.SbbIcon, HTMLSbbIconElement>('sbb-icon');
+export const SbbImage = /*@__PURE__*/createReactComponent<JSX.SbbImage, HTMLSbbImageElement>('sbb-image');
+export const SbbJourneyHeader = /*@__PURE__*/createReactComponent<JSX.SbbJourneyHeader, HTMLSbbJourneyHeaderElement>('sbb-journey-header');
+export const SbbLink = /*@__PURE__*/createReactComponent<JSX.SbbLink, HTMLSbbLinkElement>('sbb-link');
+export const SbbLinkButton = /*@__PURE__*/createReactComponent<JSX.SbbLinkButton, HTMLSbbLinkButtonElement>('sbb-link-button');
+export const SbbLinkList = /*@__PURE__*/createReactComponent<JSX.SbbLinkList, HTMLSbbLinkListElement>('sbb-link-list');
+export const SbbLogo = /*@__PURE__*/createReactComponent<JSX.SbbLogo, HTMLSbbLogoElement>('sbb-logo');
+export const SbbOverlay = /*@__PURE__*/createReactComponent<JSX.SbbOverlay, HTMLSbbOverlayElement>('sbb-overlay');
+export const SbbPanel = /*@__PURE__*/createReactComponent<JSX.SbbPanel, HTMLSbbPanelElement>('sbb-panel');
+export const SbbPearlChain = /*@__PURE__*/createReactComponent<JSX.SbbPearlChain, HTMLSbbPearlChainElement>('sbb-pearl-chain');
+export const SbbSection = /*@__PURE__*/createReactComponent<JSX.SbbSection, HTMLSbbSectionElement>('sbb-section');
+export const SbbSignet = /*@__PURE__*/createReactComponent<JSX.SbbSignet, HTMLSbbSignetElement>('sbb-signet');
+export const SbbSlotComponent = /*@__PURE__*/createReactComponent<JSX.SbbSlotComponent, HTMLSbbSlotComponentElement>('sbb-slot-component');
+export const SbbStack = /*@__PURE__*/createReactComponent<JSX.SbbStack, HTMLSbbStackElement>('sbb-stack');
+export const SbbTabAmount = /*@__PURE__*/createReactComponent<JSX.SbbTabAmount, HTMLSbbTabAmountElement>('sbb-tab-amount');
+export const SbbTabGroup = /*@__PURE__*/createReactComponent<JSX.SbbTabGroup, HTMLSbbTabGroupElement>('sbb-tab-group');
+export const SbbTabTitle = /*@__PURE__*/createReactComponent<JSX.SbbTabTitle, HTMLSbbTabTitleElement>('sbb-tab-title');
+export const SbbTeaser = /*@__PURE__*/createReactComponent<JSX.SbbTeaser, HTMLSbbTeaserElement>('sbb-teaser');
+export const SbbTeaserHero = /*@__PURE__*/createReactComponent<JSX.SbbTeaserHero, HTMLSbbTeaserHeroElement>('sbb-teaser-hero');
+export const SbbTimetable = /*@__PURE__*/createReactComponent<JSX.SbbTimetable, HTMLSbbTimetableElement>('sbb-timetable');
+export const SbbTimetableBarrierFree = /*@__PURE__*/createReactComponent<JSX.SbbTimetableBarrierFree, HTMLSbbTimetableBarrierFreeElement>('sbb-timetable-barrier-free');
+export const SbbTimetableButton = /*@__PURE__*/createReactComponent<JSX.SbbTimetableButton, HTMLSbbTimetableButtonElement>('sbb-timetable-button');
+export const SbbTimetableCusHim = /*@__PURE__*/createReactComponent<JSX.SbbTimetableCusHim, HTMLSbbTimetableCusHimElement>('sbb-timetable-cus-him');
+export const SbbTimetableDuration = /*@__PURE__*/createReactComponent<JSX.SbbTimetableDuration, HTMLSbbTimetableDurationElement>('sbb-timetable-duration');
+export const SbbTimetableOccupancy = /*@__PURE__*/createReactComponent<JSX.SbbTimetableOccupancy, HTMLSbbTimetableOccupancyElement>('sbb-timetable-occupancy');
+export const SbbTimetableParkAndRail = /*@__PURE__*/createReactComponent<JSX.SbbTimetableParkAndRail, HTMLSbbTimetableParkAndRailElement>('sbb-timetable-park-and-rail');
+export const SbbTimetablePlatform = /*@__PURE__*/createReactComponent<JSX.SbbTimetablePlatform, HTMLSbbTimetablePlatformElement>('sbb-timetable-platform');
+export const SbbTimetableRow = /*@__PURE__*/createReactComponent<JSX.SbbTimetableRow, HTMLSbbTimetableRowElement>('sbb-timetable-row');
+export const SbbTimetableRowButton = /*@__PURE__*/createReactComponent<JSX.SbbTimetableRowButton, HTMLSbbTimetableRowButtonElement>('sbb-timetable-row-button');
+export const SbbTimetableRowColumnHeaders = /*@__PURE__*/createReactComponent<JSX.SbbTimetableRowColumnHeaders, HTMLSbbTimetableRowColumnHeadersElement>('sbb-timetable-row-column-headers');
+export const SbbTimetableRowDayChange = /*@__PURE__*/createReactComponent<JSX.SbbTimetableRowDayChange, HTMLSbbTimetableRowDayChangeElement>('sbb-timetable-row-day-change');
+export const SbbTimetableRowHeader = /*@__PURE__*/createReactComponent<JSX.SbbTimetableRowHeader, HTMLSbbTimetableRowHeaderElement>('sbb-timetable-row-header');
+export const SbbTimetableTransportationDetails = /*@__PURE__*/createReactComponent<JSX.SbbTimetableTransportationDetails, HTMLSbbTimetableTransportationDetailsElement>('sbb-timetable-transportation-details');
+export const SbbTimetableTransportationNumber = /*@__PURE__*/createReactComponent<JSX.SbbTimetableTransportationNumber, HTMLSbbTimetableTransportationNumberElement>('sbb-timetable-transportation-number');
+export const SbbTimetableTransportationTime = /*@__PURE__*/createReactComponent<JSX.SbbTimetableTransportationTime, HTMLSbbTimetableTransportationTimeElement>('sbb-timetable-transportation-time');
+export const SbbTimetableTransportationWalk = /*@__PURE__*/createReactComponent<JSX.SbbTimetableTransportationWalk, HTMLSbbTimetableTransportationWalkElement>('sbb-timetable-transportation-walk');
+export const SbbTimetableTravelHints = /*@__PURE__*/createReactComponent<JSX.SbbTimetableTravelHints, HTMLSbbTimetableTravelHintsElement>('sbb-timetable-travel-hints');
+export const SbbTitle = /*@__PURE__*/createReactComponent<JSX.SbbTitle, HTMLSbbTitleElement>('sbb-title');
+export const SbbToggleCheck = /*@__PURE__*/createReactComponent<JSX.SbbToggleCheck, HTMLSbbToggleCheckElement>('sbb-toggle-check');