+++ src/components.d.ts
@@ -12,10 +12,12 @@ import { InterfaceCardProductAttributes } from "./components/sbb-card-product/sb
 import { Time } from "./components/sbb-clock/sbb-clock.custom";
 import { InterfaceSbbDividerAttributes } from "./components/sbb-divider/sbb-divider.custom.d";
 import { InterfaceFooterAttributes } from "./components/sbb-footer/sbb-footer.custom";
+import { InterfaceSbbFormFieldAttributes } from "./components/sbb-form-field/sbb-form-field.custom";
 import { InterfaceGridAttributes } from "./components/sbb-grid/sbb-grid.custom";
 import { InterfaceImageAttributes } from "./components/sbb-image/sbb-image.custom";
 import { InterfaceJourneyHeaderAttributes } from "./components/sbb-journey-header/sbb-journey-header.custom";
 import { InterfaceLinkAttributes } from "./components/sbb-link/sbb-link.custom";
+import { LinkTargetType } from "./global/interfaces/link-button-properties";
 import { InterfaceLinkButtonAttributes } from "./components/sbb-link-button/sbb-link-button.custom";
 import { InterfaceTitleAttributes } from "./components/sbb-title/sbb-title.custom.d";
 import { InterfaceLinkListAttributes } from "./components/sbb-link-list/sbb-link-list.custom";
@@ -28,7 +30,6 @@ import { InterfaceSignetAttributes } from "./components/sbb-signet/sbb-signet.cu
 import { InterfaceStackAttributes } from "./components/sbb-stack/sbb-stack.custom";
 import { InterfaceTabTitleAttributes } from "./components/sbb-tab-title/sbb-tab-title.custom";
 import { InterfaceTeaserAttributes } from "./components/sbb-teaser/sbb-teaser.custom";
-import { InterfaceTextInputAttributes } from "./components/sbb-text-input/sbb-text-input.custom";
 import { InterfaceTimetableButtonAttributes } from "./components/sbb-timetable-button/sbb-timetable-button.custom";
 import { InterfaceTimetableCusHimAttributes } from "./components/sbb-timetable-cus-him/sbb-timetable-cus-him.custom";
 import { InterfaceTimetableParkAndRailAttributes } from "./components/sbb-timetable-park-and-rail/sbb-timetable-park-and-rail.custom";
@@ -316,6 +317,30 @@ export namespace Components {
          */
         "appearance"?: InterfaceFooterAttributes['appearance'];
     }
+    interface SbbFormError {
+    }
+    interface SbbFormField {
+        /**
+          * Whether to display the form field without a border.
+         */
+        "borderless": boolean;
+        /**
+          * Whether to reserve space for an error message. `none` does not reserve any space. `reserve` does reserve one row for an error message.
+         */
+        "errorSpace"?: InterfaceSbbFormFieldAttributes['errorSpace'];
+        /**
+          * Label text for the input which is internally rendered as `<label>`.
+         */
+        "label": string;
+        /**
+          * Indicates whether the input is optional.
+         */
+        "optional"?: boolean;
+        /**
+          * Size variant, either l or m.
+         */
+        "size"?: InterfaceSbbFormFieldAttributes['size'];
+    }
     interface SbbGrid {
         /**
           * Section appearance
@@ -415,12 +440,6 @@ export namespace Components {
          */
         "pictureSizesConfig"?: string;
     }
-    interface SbbInputError {
-        /**
-          * The error message, we want to show.
-         */
-        "message": string;
-    }
     interface SbbJourneyHeader {
         /**
           * Journey header appearance
@@ -505,6 +524,14 @@ export namespace Components {
          */
         "negative": boolean;
         /**
+          * The relationship of the linked URL as space-separated link types.
+         */
+        "rel"?: string | undefined;
+        /**
+          * Where to display the linked URL.
+         */
+        "target"?: LinkTargetType | string | undefined;
+        /**
           * Text size, the link should get in the non-button variation. With inline variant, the text size adapts to where it is used.
          */
         "textSize": InterfaceLinkAttributes['textSize'];
@@ -512,6 +539,10 @@ export namespace Components {
           * Type attribute if link is used as button (optional)
          */
         "type": InterfaceLinkAttributes['buttonType'] | undefined;
+        /**
+          * The value associated with button `name` when it's submitted with the form data.
+         */
+        "value"?: string;
         /**
           * Applies link inline styles (underline, inherit coloring/font-size etc).
          */
@@ -790,96 +821,6 @@ export namespace Components {
           * Text property for sbb-panel. See sbb-panel for additional info
          */
         "text": string;
-    }
-    interface SbbTextInput {
-        /**
-          * If set to true, the input element will have no border, but a drop shadow.
-         */
-        "borderless"?: boolean;
-        /**
-          * Debounce type for the input change event in ms. If you set this value to e.g. 300, we fire the input event only every 300ms.
-         */
-        "debounceInputEvent"?: number;
-        /**
-          * Id which is sent as the id in the eventDetail payload
-         */
-        "eventId"?: string;
-        /**
-          * The icon name we want to use, choose from the small icon variants from the ui-icons category from here https://lyne.sbb.ch/tokens/icons/.
-         */
-        "icon"?: string;
-        /**
-          * The aria-autocomplete attribute for the input element.
-         */
-        "inputAriaAutoComplete"?: InterfaceTextInputAttributes['inputAriaAutoComplete'];
-        /**
-          * The id to use as the aira-controls attribute for the input element.
-         */
-        "inputAriaControls"?: string;
-        /**
-          * Set aria-expanded on the input element.
-         */
-        "inputAriaExpanded"?: boolean;
-        /**
-          * Choose either on, off or one of the existing autocomplete values. Read more about them here: https://mzl.la/3wpfaDV
-         */
-        "inputAutoCompleteValue"?: InterfaceTextInputAttributes['inputAutoCompleteValue'];
-        /**
-          * If set to true, the input field will be disabled.
-         */
-        "inputDisabled"?: boolean;
-        /**
-          * If set to true, we will set an an error message for the current input field.
-         */
-        "inputError"?: boolean;
-        /**
-          * Each input needs to have an individual id. If no id is provided, the component will create a unique id by itself.
-         */
-        "inputId"?: string;
-        /**
-          * Pass on a expected max length.
-         */
-        "inputMaxLength"?: number;
-        /**
-          * Pass on a expected min length.
-         */
-        "inputMinLength"?: number;
-        /**
-          * Each input should have an individual name.
-         */
-        "inputName": string;
-        /**
-          * Add a validation pattern (regex) the input should follow. Read more here: https://mzl.la/3C3HTiG
-         */
-        "inputPattern"?: string;
-        /**
-          * Add a placeholder to show what kind of input is expected.
-         */
-        "inputPlaceholder"?: string;
-        /**
-          * If set to true, an input in this field will be required.
-         */
-        "inputRequired"?: boolean;
-        /**
-          * The role attribute used for the input element.
-         */
-        "inputRole"?: InterfaceTextInputAttributes['inputRole'];
-        /**
-          * Define which input type you would like to use. Read more about the individual advantages here, most of the are related to show the user the most convienient keyboard: https://bit.ly/3wuQE47
-         */
-        "inputType": string;
-        /**
-          * Value for the input element.
-         */
-        "inputValue"?: string;
-        /**
-          * Each input element needs to have a label associated with it.
-         */
-        "label": string;
-        /**
-          * If set to false, the label will be visually hidden but still be in the markup to provide proper semantics
-         */
-        "labelVisible"?: boolean;
     }
     interface SbbTimetable {
     }
@@ -1188,6 +1129,18 @@ declare global {
         prototype: HTMLSbbFooterElement;
         new (): HTMLSbbFooterElement;
     };
+    interface HTMLSbbFormErrorElement extends Components.SbbFormError, HTMLStencilElement {
+    }
+    var HTMLSbbFormErrorElement: {
+        prototype: HTMLSbbFormErrorElement;
+        new (): HTMLSbbFormErrorElement;
+    };
+    interface HTMLSbbFormFieldElement extends Components.SbbFormField, HTMLStencilElement {
+    }
+    var HTMLSbbFormFieldElement: {
+        prototype: HTMLSbbFormFieldElement;
+        new (): HTMLSbbFormFieldElement;
+    };
     interface HTMLSbbGridElement extends Components.SbbGrid, HTMLStencilElement {
     }
     var HTMLSbbGridElement: {
@@ -1205,12 +1158,6 @@ declare global {
     var HTMLSbbImageElement: {
         prototype: HTMLSbbImageElement;
         new (): HTMLSbbImageElement;
-    };
-    interface HTMLSbbInputErrorElement extends Components.SbbInputError, HTMLStencilElement {
-    }
-    var HTMLSbbInputErrorElement: {
-        prototype: HTMLSbbInputErrorElement;
-        new (): HTMLSbbInputErrorElement;
     };
     interface HTMLSbbJourneyHeaderElement extends Components.SbbJourneyHeader, HTMLStencilElement {
     }
@@ -1313,12 +1260,6 @@ declare global {
     var HTMLSbbTeaserHeroElement: {
         prototype: HTMLSbbTeaserHeroElement;
         new (): HTMLSbbTeaserHeroElement;
-    };
-    interface HTMLSbbTextInputElement extends Components.SbbTextInput, HTMLStencilElement {
-    }
-    var HTMLSbbTextInputElement: {
-        prototype: HTMLSbbTextInputElement;
-        new (): HTMLSbbTextInputElement;
     };
     interface HTMLSbbTimetableElement extends Components.SbbTimetable, HTMLStencilElement {
     }
@@ -1451,10 +1392,11 @@ declare global {
         "sbb-clock": HTMLSbbClockElement;
         "sbb-divider": HTMLSbbDividerElement;
         "sbb-footer": HTMLSbbFooterElement;
+        "sbb-form-error": HTMLSbbFormErrorElement;
+        "sbb-form-field": HTMLSbbFormFieldElement;
         "sbb-grid": HTMLSbbGridElement;
         "sbb-icon": HTMLSbbIconElement;
         "sbb-image": HTMLSbbImageElement;
-        "sbb-input-error": HTMLSbbInputErrorElement;
         "sbb-journey-header": HTMLSbbJourneyHeaderElement;
         "sbb-link": HTMLSbbLinkElement;
         "sbb-link-button": HTMLSbbLinkButtonElement;
@@ -1472,7 +1414,6 @@ declare global {
         "sbb-tab-title": HTMLSbbTabTitleElement;
         "sbb-teaser": HTMLSbbTeaserElement;
         "sbb-teaser-hero": HTMLSbbTeaserHeroElement;
-        "sbb-text-input": HTMLSbbTextInputElement;
         "sbb-timetable": HTMLSbbTimetableElement;
         "sbb-timetable-barrier-free": HTMLSbbTimetableBarrierFreeElement;
         "sbb-timetable-button": HTMLSbbTimetableButtonElement;
@@ -1776,6 +1717,30 @@ declare namespace LocalJSX {
          */
         "appearance"?: InterfaceFooterAttributes['appearance'];
     }
+    interface SbbFormError {
+    }
+    interface SbbFormField {
+        /**
+          * Whether to display the form field without a border.
+         */
+        "borderless"?: boolean;
+        /**
+          * Whether to reserve space for an error message. `none` does not reserve any space. `reserve` does reserve one row for an error message.
+         */
+        "errorSpace"?: InterfaceSbbFormFieldAttributes['errorSpace'];
+        /**
+          * Label text for the input which is internally rendered as `<label>`.
+         */
+        "label"?: string;
+        /**
+          * Indicates whether the input is optional.
+         */
+        "optional"?: boolean;
+        /**
+          * Size variant, either l or m.
+         */
+        "size"?: InterfaceSbbFormFieldAttributes['size'];
+    }
     interface SbbGrid {
         /**
           * Section appearance
@@ -1875,12 +1840,6 @@ declare namespace LocalJSX {
          */
         "pictureSizesConfig"?: string;
     }
-    interface SbbInputError {
-        /**
-          * The error message, we want to show.
-         */
-        "message": string;
-    }
     interface SbbJourneyHeader {
         /**
           * Journey header appearance
@@ -1969,6 +1928,14 @@ declare namespace LocalJSX {
          */
         "onSbb-link-button_click"?: (event: SbbLinkCustomEvent<any>) => void;
         /**
+          * The relationship of the linked URL as space-separated link types.
+         */
+        "rel"?: string | undefined;
+        /**
+          * Where to display the linked URL.
+         */
+        "target"?: LinkTargetType | string | undefined;
+        /**
           * Text size, the link should get in the non-button variation. With inline variant, the text size adapts to where it is used.
          */
         "textSize"?: InterfaceLinkAttributes['textSize'];
@@ -1976,6 +1943,10 @@ declare namespace LocalJSX {
           * Type attribute if link is used as button (optional)
          */
         "type"?: InterfaceLinkAttributes['buttonType'] | undefined;
+        /**
+          * The value associated with button `name` when it's submitted with the form data.
+         */
+        "value"?: string;
         /**
           * Applies link inline styles (underline, inherit coloring/font-size etc).
          */
@@ -2240,96 +2211,6 @@ declare namespace LocalJSX {
          */
         "text": string;
     }
-    interface SbbTextInput {
-        /**
-          * If set to true, the input element will have no border, but a drop shadow.
-         */
-        "borderless"?: boolean;
-        /**
-          * Debounce type for the input change event in ms. If you set this value to e.g. 300, we fire the input event only every 300ms.
-         */
-        "debounceInputEvent"?: number;
-        /**
-          * Id which is sent as the id in the eventDetail payload
-         */
-        "eventId"?: string;
-        /**
-          * The icon name we want to use, choose from the small icon variants from the ui-icons category from here https://lyne.sbb.ch/tokens/icons/.
-         */
-        "icon"?: string;
-        /**
-          * The aria-autocomplete attribute for the input element.
-         */
-        "inputAriaAutoComplete"?: InterfaceTextInputAttributes['inputAriaAutoComplete'];
-        /**
-          * The id to use as the aira-controls attribute for the input element.
-         */
-        "inputAriaControls"?: string;
-        /**
-          * Set aria-expanded on the input element.
-         */
-        "inputAriaExpanded"?: boolean;
-        /**
-          * Choose either on, off or one of the existing autocomplete values. Read more about them here: https://mzl.la/3wpfaDV
-         */
-        "inputAutoCompleteValue"?: InterfaceTextInputAttributes['inputAutoCompleteValue'];
-        /**
-          * If set to true, the input field will be disabled.
-         */
-        "inputDisabled"?: boolean;
-        /**
-          * If set to true, we will set an an error message for the current input field.
-         */
-        "inputError"?: boolean;
-        /**
-          * Each input needs to have an individual id. If no id is provided, the component will create a unique id by itself.
-         */
-        "inputId"?: string;
-        /**
-          * Pass on a expected max length.
-         */
-        "inputMaxLength"?: number;
-        /**
-          * Pass on a expected min length.
-         */
-        "inputMinLength"?: number;
-        /**
-          * Each input should have an individual name.
-         */
-        "inputName": string;
-        /**
-          * Add a validation pattern (regex) the input should follow. Read more here: https://mzl.la/3C3HTiG
-         */
-        "inputPattern"?: string;
-        /**
-          * Add a placeholder to show what kind of input is expected.
-         */
-        "inputPlaceholder"?: string;
-        /**
-          * If set to true, an input in this field will be required.
-         */
-        "inputRequired"?: boolean;
-        /**
-          * The role attribute used for the input element.
-         */
-        "inputRole"?: InterfaceTextInputAttributes['inputRole'];
-        /**
-          * Define which input type you would like to use. Read more about the individual advantages here, most of the are related to show the user the most convienient keyboard: https://bit.ly/3wuQE47
-         */
-        "inputType": string;
-        /**
-          * Value for the input element.
-         */
-        "inputValue"?: string;
-        /**
-          * Each input element needs to have a label associated with it.
-         */
-        "label": string;
-        /**
-          * If set to false, the label will be visually hidden but still be in the markup to provide proper semantics
-         */
-        "labelVisible"?: boolean;
-    }
     interface SbbTimetable {
     }
     interface SbbTimetableBarrierFree {
@@ -2570,10 +2451,11 @@ declare namespace LocalJSX {
         "sbb-clock": SbbClock;
         "sbb-divider": SbbDivider;
         "sbb-footer": SbbFooter;
+        "sbb-form-error": SbbFormError;
+        "sbb-form-field": SbbFormField;
         "sbb-grid": SbbGrid;
         "sbb-icon": SbbIcon;
         "sbb-image": SbbImage;
-        "sbb-input-error": SbbInputError;
         "sbb-journey-header": SbbJourneyHeader;
         "sbb-link": SbbLink;
         "sbb-link-button": SbbLinkButton;
@@ -2591,7 +2473,6 @@ declare namespace LocalJSX {
         "sbb-tab-title": SbbTabTitle;
         "sbb-teaser": SbbTeaser;
         "sbb-teaser-hero": SbbTeaserHero;
-        "sbb-text-input": SbbTextInput;
         "sbb-timetable": SbbTimetable;
         "sbb-timetable-barrier-free": SbbTimetableBarrierFree;
         "sbb-timetable-button": SbbTimetableButton;
@@ -2628,10 +2509,11 @@ declare module "@stencil/core" {
             "sbb-clock": LocalJSX.SbbClock & JSXBase.HTMLAttributes<HTMLSbbClockElement>;
             "sbb-divider": LocalJSX.SbbDivider & JSXBase.HTMLAttributes<HTMLSbbDividerElement>;
             "sbb-footer": LocalJSX.SbbFooter & JSXBase.HTMLAttributes<HTMLSbbFooterElement>;
+            "sbb-form-error": LocalJSX.SbbFormError & JSXBase.HTMLAttributes<HTMLSbbFormErrorElement>;
+            "sbb-form-field": LocalJSX.SbbFormField & JSXBase.HTMLAttributes<HTMLSbbFormFieldElement>;
             "sbb-grid": LocalJSX.SbbGrid & JSXBase.HTMLAttributes<HTMLSbbGridElement>;
             "sbb-icon": LocalJSX.SbbIcon & JSXBase.HTMLAttributes<HTMLSbbIconElement>;
             "sbb-image": LocalJSX.SbbImage & JSXBase.HTMLAttributes<HTMLSbbImageElement>;
-            "sbb-input-error": LocalJSX.SbbInputError & JSXBase.HTMLAttributes<HTMLSbbInputErrorElement>;
             "sbb-journey-header": LocalJSX.SbbJourneyHeader & JSXBase.HTMLAttributes<HTMLSbbJourneyHeaderElement>;
             "sbb-link": LocalJSX.SbbLink & JSXBase.HTMLAttributes<HTMLSbbLinkElement>;
             "sbb-link-button": LocalJSX.SbbLinkButton & JSXBase.HTMLAttributes<HTMLSbbLinkButtonElement>;
@@ -2649,7 +2531,6 @@ declare module "@stencil/core" {
             "sbb-tab-title": LocalJSX.SbbTabTitle & JSXBase.HTMLAttributes<HTMLSbbTabTitleElement>;
             "sbb-teaser": LocalJSX.SbbTeaser & JSXBase.HTMLAttributes<HTMLSbbTeaserElement>;
             "sbb-teaser-hero": LocalJSX.SbbTeaserHero & JSXBase.HTMLAttributes<HTMLSbbTeaserHeroElement>;
-            "sbb-text-input": LocalJSX.SbbTextInput & JSXBase.HTMLAttributes<HTMLSbbTextInputElement>;
             "sbb-timetable": LocalJSX.SbbTimetable & JSXBase.HTMLAttributes<HTMLSbbTimetableElement>;
             "sbb-timetable-barrier-free": LocalJSX.SbbTimetableBarrierFree & JSXBase.HTMLAttributes<HTMLSbbTimetableBarrierFreeElement>;
             "sbb-timetable-button": LocalJSX.SbbTimetableButton & JSXBase.HTMLAttributes<HTMLSbbTimetableButtonElement>;